Методы валидации перенесены в сервисный слой
