+++ src/main/java/ru/yandex/practicum/filmorate/service/UserService.java
@@ -4,11 +4,13 @@ import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
+import ru.yandex.practicum.filmorate.exception.BadRequestException;
 import ru.yandex.practicum.filmorate.exception.NotFoundException;
 import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.User;
-import ru.yandex.practicum.filmorate.storage.user.UserStorage;
+import ru.yandex.practicum.filmorate.storage.user.InMemoryUserStorage;
 
+import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
@@ -19,7 +21,8 @@ import java.util.Map;
 @Service
 @RequiredArgsConstructor
 public class UserService {
-    private final UserStorage userStorage;
+    private final InMemoryUserStorage userStorage;
+
     @Getter
     private final Map<Integer, List<User>> friendsList = new HashMap<>();
 
@@ -91,5 +94,51 @@ public class UserService {
                 .filter(user -> user.getId() != idTwo)
                 .toList();
     }
+
+    public User checkForCreate(User user) {
+        if (user.getLogin().contains(" ")) {
+            log.error("Пользователь ввел логин с пробелом");
+            throw new BadRequestException("Поле логин не может содержать пробелы");
+        }
+        if (isValueNull(user.getName()) || isLineBlank(user.getName())) {
+            user.setName(user.getLogin());
+        }
+        if (!isDateNull(user.getBirthday())) {
+            if (user.getBirthday().isAfter(LocalDate.now())) {
+                log.error("Пользователь ввел дату из будущего");
+                throw new BadRequestException("Дата рождения не может быть в будущем");
+            }
+        }
+        return user;
+    }
+
+    public User checkForUpdate(User user) {
+        if (isIdNull(user.getId())) {
+            log.error("Пользователь не ввел id");
+            throw new ValidationException("Id должен быть указан");
+        }
+        if (userStorage.getUsers().containsKey(user.getId())) {
+            return checkForCreate(user);
+        }
+        log.error("Фильм с= " + user.getId() + " не найден");
+        throw new ValidationException("Фильм с id = " + user.getId() + " не найден");
+    }
+
+    public boolean isIdNull(Integer id) {
+        return id == 0;
+    }
+
+    public boolean isDateNull(LocalDate date) {
+        return date == null;
+    }
+
+    public boolean isLineBlank(String value) {
+        return value.isBlank();
+    }
+
+    public boolean isValueNull(String value) {
+        return value == null;
+    }
+
 }
