+++ src/main/java/ru/yandex/practicum/filmorate/storage/film/FilmStorage.java
@@ -3,7 +3,6 @@ package ru.yandex.practicum.filmorate.storage.film;
 import org.springframework.stereotype.Component;
 import ru.yandex.practicum.filmorate.model.Film;
 
-import java.time.LocalDate;
 import java.util.List;
 import java.util.Map;
 
@@ -14,17 +13,12 @@ public interface FilmStorage {
 
     List<Film> getAllFilms();
 
+    List<Film> popularFilm(String size);
+
     int getNextId();
-
-    boolean isIdNull(Integer id);
-
-    boolean isDateNull(LocalDate date);
 
     Film createFilm(Film film);
 
     Film updateFilm(Film film);
 
-    void checkDate(Film film);
-
-    Film checkForUpdate(Film film);
 }
