+++ src/main/java/ru/yandex/practicum/filmorate/storage/user/UserStorage.java
@@ -3,7 +3,6 @@ package ru.yandex.practicum.filmorate.storage.user;
 import org.springframework.stereotype.Component;
 import ru.yandex.practicum.filmorate.model.User;
 
-import java.time.LocalDate;
 import java.util.List;
 import java.util.Map;
 
@@ -16,19 +15,8 @@ public interface UserStorage {
 
     int getNextId();
 
-    boolean isIdNull(Integer id);
-
-    boolean isDateNull(LocalDate date);
-
-    boolean isLineBlank(String value);
-
-    boolean isValueNull(String value);
-
     User createUser(User user);
 
     User updateUser(User user);
 
-    User checkForCreate(User user);
-
-    User checkForUpdate(User user);
 }
